\name{cgd1}
\alias{cgd1}
\docType{data}
\title{
cgd1 Data Set
}
\description{
This database was obtained from the survival package containing a time-to-event data.
}
\usage{data(weight_behavior)}
\format{
The data set contains many variables.
}
\examples{
data(cgd1)
names(cgd1)
}
\keyword{Datasets}
